Fix standalone horizon 1.0.0 (#164)
Horizon 1.0.0 requires history archives to build internal ledger state.
This breaks standalone mode because Stellar-Core stores history data in
local file system. To solve this we start local web server that serves
files from history archive.

This is a quick fix: when history is initialized using `stellar-core
newhist` command the `currentLedger` value is equal `0` and makes
Horizon wait for a valid, non-zero value. Because of this Horizon is
able to start ingesting after 64 ledgers (when first checkpoint is
created).

To fix this permanently Stellar-Core or Horizon code needs to be
changed.